Add files via upload
A generalized model that used the data from UCI machine learning repository to predict if a person has donated blood or not. 
This consists of features like recency, frequency, c.c of blood donated, time and predicts a binary output.
This is trained by Logistic Regression Algorithm using various libraries like pandas, keras etc.
The last part of code is not working due to some dimension errors that I'll try to fix soon.